+++ integration/js/DataMessageTest.js
@@ -4,7 +4,7 @@ const {TestUtils} = require('./node_modules/kite-common');
 const SdkBaseTest = require('./utils/SdkBaseTest');
 const {SdkTestUtils} = require('./utils/SdkTestUtils');
 const {Window} = require('./utils/Window');
-const uuidv4 = require('uuid/v4');
+const { v4: uuidv4 } = require('uuid');
 
 /*
 * 1. First participant joins the meeting
@@ -22,7 +22,7 @@ class DataMessageTest extends SdkBaseTest {
   }
 
   async runIntegrationTest() {
-    if (this.numberOfSessions(this.capabilities.browserName) > 1) {
+    if (this.numberOfSessions() > 1) {
       await this.runTestOnMultipleSessions()
     } else {
       await this.runTestOnSingleSessions()